\name{Rlateblight-package}
\alias{Rlateblight-package}
\alias{Rlateblight}

\docType{package}

\title{
Forecasters of potato late blight and potato late blight dynamic simulation model
}

\description{
Potato late blight is a plant disease caused by oomycete phytophthora infestans.
}

\author{
Robert J. Hijmans  <r.hijmans@gmail.com>
}
